+++ AppointmentSchedulerServer/BusinessLogicLayer/Interfaces/IAppointmentBLL.cs
@@ -6,8 +6,8 @@ namespace AppointmentSchedulerServer.BusinessLogicLayer.Interfaces
 {
     public interface IAppointmentBLL
     {
-        public Task<ActionResult<IEnumerable<Appointment>>> FindAllByAccountIdAsync(int id);
-        public Task<ActionResult<IEnumerable<Appointment>>> FindAllByEmployeeIdAsync(int id);
+        public Task<ActionResult<IEnumerable<Appointment>>> FindAllByAccountIdAsync(long id);
+        public Task<ActionResult<IEnumerable<Appointment>>> FindAllByEmployeeIdAsync(long id);
         public Task<ActionResult> Save(CreateAppointmentDTO appointmentDTO);
         public Task<ActionResult<IEnumerable<AppointmentTypeDTO>>> GetAllAppointmentTypes();
         public Task<ActionResult<IEnumerable<GetEmployeeDTO>>> GetEmployeeByAppointmentType(long id);
